refactor(vcs): SOLID/DRY pass — extract shared utilities, separate bootstrap from runtime
