+++ packages/service/src/api/handlers/vcs/vcsShow.ts
@@ -3,18 +3,15 @@
  * Fastify route handler for GET /vcs/show. Returns file content at a specific commit.
  */
 
-import { execFile } from 'node:child_process';
 import { extname } from 'node:path';
-import { promisify } from 'node:util';
 
 import type { FastifyReply, FastifyRequest } from 'fastify';
 import type pino from 'pino';
 
+import { execFileAsync } from '../../../vcs/gitExec';
 import { resolveWatchRoot } from '../../../vcs/resolveWatchRoot';
 import type { VcsCoordinator } from '../../../vcs/VcsCoordinator';
 import { wrapHandler } from '../wrapHandler';
-
-const execFileAsync = promisify(execFile);
 
 export interface VcsShowRouteDeps {
   coordinator: VcsCoordinator;
