+++ packages/service/src/vcs/VcsManager.ts
@@ -3,62 +3,20 @@
  * Per-root VCS manager for git-backed content versioning.
  */
 
-import { execFile } from 'node:child_process';
-import { access, appendFile, readFile, writeFile } from 'node:fs/promises';
-import { join } from 'node:path';
-import { promisify } from 'node:util';
-
 import type { VcsConfig } from '@karmaniverous/jeeves-watcher-core';
 import type pino from 'pino';
 
 import { normalizeError } from '../util/normalizeError';
+import { retry } from '../util/retry';
 import type { CommitMessageGenerator } from './CommitMessageGenerator';
+import { execFileAsync } from './gitExec';
 import { SquashManager } from './SquashManager';
-
-const execFileAsync = promisify(execFile);
 
 /** Metadata for a pending reversion to include in the commit message. */
 export interface PendingReversion {
   glob: string;
   commit: string;
   paths: string[];
-}
-
-/** Always-on .gitignore entries for VCS-managed watch roots. */
-const ALWAYS_GITIGNORE_ENTRIES = [
-  '.git/',
-  'node_modules/',
-  '.jeeves-watcher/',
-  '.jeeves-metadata/',
-];
-
-/** Maximum retries for index.lock contention. */
-const MAX_LOCK_RETRIES = 3;
-
-/** Backoff delays in ms for index.lock retries. */
-const LOCK_RETRY_DELAYS = [500, 1000, 2000];
-
-/**
- * Check whether an error is caused by index.lock contention.
- */
-function isIndexLockError(error: unknown): boolean {
-  if (!(error instanceof Error)) return false;
-  const message = error.message || '';
-  const stderr =
-    'stderr' in error ? String((error as { stderr: unknown }).stderr) : '';
-  return (
-    message.includes('index.lock') ||
-    stderr.includes('index.lock') ||
-    message.includes('EEXIST') ||
-    stderr.includes('EEXIST')
-  );
-}
-
-/**
- * Sleep for the given number of milliseconds.
- */
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
 /** Push error record for status reporting. */
@@ -125,78 +83,6 @@ export class VcsManager {
 
   get pushErrors(): readonly PushError[] {
     return this._pushErrors;
-  }
-
-  /**
-   * Check whether git is available on the system PATH.
-   *
-   * @returns true if git is available, false otherwise.
-   */
-  static async checkGitAvailable(): Promise<boolean> {
-    try {
-      await execFileAsync('git', ['--version']);
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * Initialize a git repository at the given path if none exists.
-   * Idempotent — does nothing if .git/ already exists.
-   *
-   * @param rootPath - Directory to initialize as a git repo.
-   */
-  static async initRepo(rootPath: string): Promise<void> {
-    const gitDir = join(rootPath, '.git');
-    try {
-      await access(gitDir);
-      // .git exists, nothing to do
-    } catch {
-      await execFileAsync('git', ['init'], { cwd: rootPath });
-    }
-  }
-
-  /**
-   * Ensure a .gitignore file exists at rootPath with all always-on entries.
-   * Creates the file if missing. Appends only entries not already present.
-   *
-   * @param rootPath - Directory containing the .gitignore.
-   * @param alwaysOnEntries - Additional always-on entries beyond the defaults.
-   */
-  static async ensureGitignore(
-    rootPath: string,
-    alwaysOnEntries: string[] = [],
-  ): Promise<void> {
-    const gitignorePath = join(rootPath, '.gitignore');
-    const requiredEntries = [...ALWAYS_GITIGNORE_ENTRIES, ...alwaysOnEntries];
-
-    let existingContent = '';
-    try {
-      existingContent = await readFile(gitignorePath, 'utf8');
-    } catch {
-      // File doesn't exist yet
-    }
-
-    const existingLines = new Set(
-      existingContent.split('\n').map((line) => line.trim()),
-    );
-    const missing = requiredEntries.filter(
-      (entry) => !existingLines.has(entry),
-    );
-
-    if (missing.length === 0) return;
-
-    if (existingContent.length === 0) {
-      await writeFile(gitignorePath, missing.join('\n') + '\n', 'utf8');
-    } else {
-      const prefix = existingContent.endsWith('\n') ? '' : '\n';
-      await appendFile(
-        gitignorePath,
-        prefix + missing.join('\n') + '\n',
-        'utf8',
-      );
-    }
   }
 
   /**
@@ -408,50 +294,51 @@ export class VcsManager {
   private async commitBatch(files: string[]): Promise<void> {
     if (files.length === 0) return;
 
-    for (let attempt = 0; attempt <= MAX_LOCK_RETRIES; attempt++) {
-      try {
-        await execFileAsync('git', ['add', '--', ...files], {
-          cwd: this.rootPath,
-        });
+    try {
+      await retry(
+        async (attempt) => {
+          await execFileAsync('git', ['add', '--', ...files], {
+            cwd: this.rootPath,
+          });
 
-        // Build message after staging so getStagedDiff can see the changes
-        const message =
-          attempt === 0
-            ? await this.buildCommitMessage(files)
-            : this.buildTemplateMessage(files.length);
-        this.pendingReversions.length = 0;
+          // Build message after staging so getStagedDiff can see the changes
+          const message =
+            attempt === 1
+              ? await this.buildCommitMessage(files)
+              : this.buildTemplateMessage(files.length);
+          this.pendingReversions.length = 0;
 
-        const { stdout } = await execFileAsync(
-          'git',
-          ['commit', '-m', message],
-          { cwd: this.rootPath },
-        );
-
-        const hashMatch = /\[[\w-]+ ([a-f0-9]+)\]/.exec(stdout);
-        const hash = hashMatch?.[1] ?? 'unknown';
-        this.logger.info(
-          { root: this.rootPath, hash, fileCount: files.length },
-          'VCS commit created',
-        );
-        await this.pushIfConfigured();
-        return;
-      } catch (error) {
-        if (isIndexLockError(error) && attempt < MAX_LOCK_RETRIES) {
-          const delay = LOCK_RETRY_DELAYS[attempt];
-          this.logger.warn(
-            { root: this.rootPath, attempt: attempt + 1, delay },
-            'index.lock contention, retrying',
+          const { stdout } = await execFileAsync(
+            'git',
+            ['commit', '-m', message],
+            { cwd: this.rootPath },
           );
-          await sleep(delay);
-          continue;
-        }
 
-        this.logger.error(
-          { root: this.rootPath, err: normalizeError(error) },
-          'VCS commit failed',
-        );
-        return;
-      }
+          const hashMatch = /\[[\w-]+ ([a-f0-9]+)\]/.exec(stdout);
+          const hash = hashMatch?.[1] ?? 'unknown';
+          this.logger.info(
+            { root: this.rootPath, hash, fileCount: files.length },
+            'VCS commit created',
+          );
+          await this.pushIfConfigured();
+        },
+        {
+          attempts: 4,
+          baseDelayMs: 500,
+          maxDelayMs: 2000,
+          onRetry: ({ attempt, delayMs }) => {
+            this.logger.warn(
+              { root: this.rootPath, attempt, delay: delayMs },
+              'index.lock contention, retrying',
+            );
+          },
+        },
+      );
+    } catch (error) {
+      this.logger.error(
+        { root: this.rootPath, err: normalizeError(error) },
+        'VCS commit failed',
+      );
     }
   }
