+++ packages/service/src/vcs/index.ts
@@ -4,6 +4,8 @@
  */
 
 export { CommitMessageGenerator } from './CommitMessageGenerator.js';
+export { execFileAsync, findRootForPath, isIndexLockError } from './gitExec.js';
+export { resolveCommitMessageApiKey } from './resolveCommitMessageApiKey.js';
 export {
   type ResolvedWatchRoot,
   resolveWatchRoot,
@@ -11,5 +13,11 @@ export {
 } from './resolveWatchRoot.js';
 export { cronMatchesNow, SquashManager } from './SquashManager.js';
 export { validateStateDirOverlap } from './validateStateDirOverlap.js';
+export {
+  ALWAYS_GITIGNORE_ENTRIES,
+  checkGitAvailable,
+  ensureGitignore,
+  initRepo,
+} from './vcsBootstrap.js';
 export { VcsCoordinator } from './VcsCoordinator.js';
 export { VcsManager } from './VcsManager.js';
